added example tests and code cleanup
