+++ community-modules/styles/src/internal/base/parts/_calculated-columns.scss
@@ -71,4 +71,15 @@
         flex-direction: column;
         gap: calc(var(--ag-grid-size) / 2);
     }
+
+    .ag-calculated-column-actions {
+        display: flex;
+        justify-content: flex-end;
+        gap: calc(var(--ag-grid-size) / 2);
+        margin-top: var(--ag-grid-size);
+    }
+
+    .ag-calculated-column-action {
+        min-width: 72px;
+    }
 }
